feat: Allow dynamic search time range
Introduces the ability to specify the search time range in days, defaulting to 3. This change updates the frontend to include a selection for search days and modifies the backend API to accept and use this parameter in the SQL query. The HAF service is also updated to pass the search days value to the API.