feat: Conditionally show custom date range
The "Custom Date Range" option in the time range selector is now only visible when the search scope is set to "user". This improves usability by preventing users from being stuck in custom mode when switching to global search. A `useEffect` hook has been added to reset the time range to "3 days" if the scope changes to "global" while the time range is set to "custom".